Split target_analysis.py to meet line limit
- Split target_analysis.py (344→144 lines) by extracting TargetAnalyzer
- Created target_analyzer.py (180 lines) for target analysis logic
- Updated target_analysis.py to contain only enums and TargetProfile
- Progress: 2 modules still exceed 300-line limit, continuing refactoring
